// src/core/enums/reject-expert-status.enums.ts
export enum RejectExpertStatusEnum {
  Unknown = 0,
  Reject = 1,
  Fix = 2,
  Accept = 3,
  SendToCurt = 4,
  CourtResult = 5,
  SetPlans = 6,
}

// public enum RejectExpertStatusEnum
// {
//     Unknown = 0,

//     [Description("رد کارشناسی")]
//     Reject = 1,

//     [Description("حل مشکل توسط کارشناس")]
//     Fix = 2,

//     [Description("تایید")]
//     Accept = 3,

//     [Description("ارسال به دادگاه بدوی")]
//     SendToCurt = 4,

//     [Description("ثبت نتیجه دادگاه بدوی")]
//     CourtResult = 5,
// }